comments user role validation done

+++ src/resources/comments/comments-service.js
@@ -4,14 +4,14 @@ const OffensiveWordsIncludes = require("../../validators/offensive-words-include
 class CommentsService {
   constructor() {}
 
-  // async getAllComments() {
-  //   try {
-  //     const allComments =await CommentsRepository.getAllComments();
-  //     return allComments;
-  //   } catch (err) {
-  //     return err.message;
-  //   }
-  // }
+  async getAllComments() {
+    try {
+      const allComments = await CommentsRepository.getAllComments();
+      return allComments;
+    } catch (err) {
+      return err.message;
+    }
+  }
 
   async createComment(postId, newComment) {
     try {
@@ -21,25 +21,33 @@ class CommentsService {
     }
   }
 
-  // async modifyComment(commentId, comment) {
-  //   try {
-  //     return await CommentsRepository.modifyComment(commentId, comment);
-  //   } catch (err) {
-  //     return err.message;
-  //   }
-  // }
+  async modifyComment(commentId, comment, user) {
+    try {
+      const commentToModify = await CommentsRepository.getPostById(commentId);
+      if (commentToModify.user) {
+        if (user.role === 0 || user._id.equals(commentToModify.user._id)) {
+          return await CommentsRepository.modifyComment(commentId, comment);
+        } else {
+          return "No tienes permiso";
+        }
+      }
+    } catch (err) {
+      return err.message;
+    }
+  }
 
   async deleteComment(commentId, user) {
     try {
       const commentToDelete = await CommentsRepository.getPostById(commentId);
 
-      if (user.role === 0 || user._id.equals(commentToDelete.user._id)) {
-        const deletedPost = await CommentsRepository.deleteComment(commentId);
-      } else {
-        return "No puedes modificar Post de otras personas";
+      if (commentToDelete.user) {
+        if (user.role === 0 || user._id.equals(commentToDelete.user._id)) {
+          const deletedPost = await CommentsRepository.deleteComment(commentId);
+          return deletedPost;
+        } else {
+          return "No tienes permiso";
+        }
       }
-
-      return await CommentsRepository.deleteComment(commentId);
     } catch (err) {
       return err.message;
     }
